+++ lib/handler/handler.d.ts
@@ -3,7 +3,7 @@ import { context, ContextProps, slot, SlotProps, slots, SlotsProps } from '@inne
 import { switchAsync, SwitchProps } from '@innet/switch';
 import { arrayAsync, async } from '@innet/utils';
 import { serverFn } from '../experimental/serverFn';
-import { cms, CmsProps, cookie, CookieProps, error, ErrorProps, file, FileProps, formatter, FormatterProps, header, HeaderProps, proxy, ProxyProps, redirect, RedirectProps, router, RouterProps, success, SuccessProps, validation, ValidationProps } from '../plugins';
+import { access, AccessProps, cms, CmsProps, cookie, CookieProps, error, ErrorProps, file, FileProps, formatter, FormatterProps, header, HeaderProps, proxy, ProxyProps, redirect, RedirectProps, router, RouterProps, success, SuccessProps, validation, ValidationProps } from '../plugins';
 import { server, ServerProps } from '../server';
 export declare const arrayPlugins: (typeof arrayAsync)[];
 export declare const JSXPlugins: {
@@ -24,6 +24,7 @@ export declare const JSXPlugins: {
     context: typeof context;
     slot: typeof slot;
     slots: typeof slots;
+    access: typeof access;
 };
 export declare const fnPlugins: (typeof serverFn)[];
 export declare const objectPlugins: ((handler: any) => import("innet").PluginHandler)[];
@@ -48,6 +49,7 @@ declare global {
             context: ContextProps;
             slot: SlotProps;
             slots: SlotsProps;
+            access: AccessProps;
         }
     }
 }