+++ lib/plugins/validation/validation.d.ts
@@ -10,7 +10,7 @@ export interface ValidationJsxElement<T> {
     children?: any;
 }
 export interface ValidationContext {
-    handleError: (e: ValidationResponse<any>) => any;
+    handleError?: (e: ValidationResponse<any>) => any;
 }
 export declare const validationContext: Context<ValidationContext, ValidationContext>;
 export declare function validation<T extends object, E extends object>({ props, children }: ValidationJsxElement<T>, handler: any): Promise<any>;